decimal decoder
